+++ internal/round.go
@@ -1,7 +1,7 @@
 package internal
 
-func getPlayersOfRound(round []Match) []uint8 {
-	players := make([]uint8, 0, 8)
+func getPlayersOfRound(round []Match) []int {
+	players := make([]int, 0, 8)
 	for _, match := range round {
 		for _, p := range match.getPlayers() {
 			if !isInSlice(p, players) {
